peblo-backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
pypdf
google-genai
supabase
